fix pipeline issues
