+++ test.py
@@ -12,7 +12,7 @@ from tracex.extraction.prototype import function_calls as fc
 from tracex.extraction.prototype import prompts as p
 from tracex.extraction.prototype import create_xes as x
 
-text = open(u.input_path / "journey_synth_covid_0.txt").read()
+""" text = open(u.input_path / "journey_synth_covid_0.txt").read()
 df = ih.convert_text_to_bulletpoints(text)
 print(df)
 df = ih.add_start_dates(text, df)
@@ -27,19 +27,6 @@ df = ih.add_locations(df)
 print(df)
 
 ih.convert_dataframe_to_csv(df)
-x.create_xes(u.output_path / "single_trace.csv", "test", "event_information")
+x.create_xes(u.output_path / "single_trace.csv", "test", "event_information") """
 
-# output = 'The end date for the bulletpoint "Experiencing mild cough and fatigue" and the start date 20220108T0000 is 20230911T0000.'
-
-# fc_message = [
-#     {"role": "system", "content": p.END_DATE_CONTEXT},
-#     {
-#         "role": "user",
-#         "content": p.END_DATE_FUNCTION_CALL
-#         + "The text: "
-#         + output
-#     },
-# ]
-
-
-# print(u.query_gpt(fc_message, tool_choice={ "type": "function", "function": {"name": "add_end_dates"}}))
+oh.get_output(u.output_path / "single_trace.csv")
