--- tracex/extraction/content/outputs/dataframe.csv
@@ -1,7 +1,0 @@
-,event_information,start_date,end_date
-0,Experiencing mild symptoms,20220601T0000,20220608T0000
-1,Developing high fever,"June 17th, 2022",20220627T0000
-2,Experiencing difficulty breathing,"June 17th, 2022","June 20th, 2022"
-3,Getting tested for Covid-19,"June 17th, 2022",20220617T0000
-4,Isolating from family,"June 17th, 2022",20220620T0000
-5,Receiving confirmation of Covid-19 contraction,20220620T0000,20220620T0000
